Allow code to be linked in RAM and run there.
This moves the interrupt handler code into RAM, eliminating potential
stalls while waiting for Flash.

ARM code, as generated by gcc, has a property that makes me suspicious
of the Flash prefetcher: it tends to get constants by PC-relative load.
This means the Flash execution pattern doesn't actually reflect the
shape of the code.

Easier to reason about if I just put it all in SRAM.  Not like I'm
hurting for SRAM, after all!

This is not complete yet.  Currently the interrupt handlers call
library functions, such as GPIO manipulation and even copy_words,
that are not in RAM.  Will fix.