Add a 1bpp palette rasterizer.
This is a pixel format converter that takes a 1bpp framebuffer and
decodes it on the fly into 8-bit pixels, using a two-entry lookup
table.  Empirically, even with an interrupt from the horizontal
timing system, this comes out to less than 10 nanoseconds per pixel.

It leans pretty hard on the SIMD instruction set to do its work.
I haven't thoroughly evaluated the obvious alternatives, such as a
256x8 lookup table for converting pixels in batches.  Managing such
a lookup table during mode changes seems error-prone, so if I can do
it with brute force, all the better.