Add an 80x37 text mode.
The approach is similar to the 1bpp rasterizer, but with a level of
indirection between the framebuffer and the pixels, and a CLUT that
can change at every character position.

It runs about 50% slower, rasterizing an 800-pixel line in 12.1 us,
or around 15ns per pixel -- still acceptable.

The font is salvaged from last year's VGA demo.