vga/graphics_1: optimize for lines of known color.
draw_line used to take a bool parameter specifying whether to set or
clear pixels, and pass that around.  At least in my current use, the
caller knows whether it wants to set or clear statically, so I'm
increasing register pressure unnecessarily and hampering the optimizer.

Now, set_line and clear_line are the true primitives.  They specialize
the internals of the Bresenham implementation statically based on the
target state of the pixels.  draw_line, when required, now costs one
conditional branch.

This takes 1.4ms (6.9%) off the rook renderer.