vga: absorb jitter ahead of latency-sensitive IRQ.
This has been counterintuitive for the folks I've run it past -- it
certainly was not obvious to me.

Several of my demos have had intermittent, difficult-to-pin jitter
problems.  This usually manifests as noise that looks like a monitor
phase mismatch, but actually results from shifts in pixel timing.
The shifts were usually rare enough that I had a hard time catching
them on a scope -- maybe one line out of 600, and my little Tek only
syncs to NTSC, not VGA, line numbers.

In working on a new demo, I managed to write a reduction for the bug.
The interesting thing about the reduction is that it's *stable:* it
isn't gaussian- or poisson-distributed jitter, but fixed offsets in
time from where I expected events to occur.  Fixed offsets of, say,
6 cycles.

Which happens to be my Flash cycle time.

The problem became clear quickly.  Prompt exception entry on a v7M
processor requires prompt stacking of registers, which always happens
through the CPU's D-code port.  Any operation that locks that port
for more than one cycle may have a negative impact on exception
latency.  Most of the slow things are out on the APB (and thus on
the S-port), but there are two multi-cycle operations available on
the D-port:

 1. Loading data from Flash.  At 160MHz this takes six cycles on a
    cache miss, and the Flash unit ignores PLD.  (This assumes
    aligned access.)

 2. Unaligned access to any low-mapped SRAM or Flash, which (for
    reads) can lock the bus for twice as long as normal (2-12c).

(Note that load-store multiple instructions don't count as multi-
cycle, because they can be interrupted and resumed.  Interestingly,
while this isn't documented, it also appears that vldm/vstm are
interruptible, though it isn't clear if they're resumed or restarted.)

Clearly, I could limit use of Flash to the first part of hblank, and
disable unaligned accesses.  But I dislike this on two fronts:

 1. It seems to impose unreasonable structuring requirements on the
    background, best-effort code.

 2. The current behavior is effectively a denial-of-service attack
    on the latency-sensitive tasks by best-effort code, stemming out
    of a totally reasonable execution pattern.

Better to prevent the attack entirely, even at some (small) cost.

In this commit I've built a 'shock absorber' for jitter.

 - By running a second timer in lock-step with the horizontal timer,
   I get a second prioritizable horizontal timing interrupt.

 - I configure this interrupt to fire slightly before the latency
   sensitive release time.  (The precise definition of 'slightly'
   took some empirical fiddling.)

 - I prioritize it just below the latency-sensitive interrupts, so
   that they can reliably preempt it but nothing else can.

 - In the handler itself, I idle the CPU, and thus the D-code port.

This costs about 300ns of CPU time and one interrupt priority level
(bringing usage to 3/16).

In tests with instrumentation added, the shock absorber interrupt
displays the entry jitter I was seeing before, but the latency-
sensitive video interrupt that preempts it is rock solid -- just as
the microarchitecture docs would predict.

Among other things, this means I can safely scan images directly
out of Flash.