vga: fix persistent off-by-one scanline bug.
Variations of this have been with me since I started this project.  It's
nuanced, since the system is dealing with two lines of video at any
given time.  But I've done a more rigorous analysis and worked out a way
to express the state transitions that (1) fixes the bug and (2) reads
better, at least to me.

I also noticed and removed some unnecessary control flow in the
rasterizer fast path, and eliminated an unnecessary scanout copy in the
last blank line before video.  So that should improve performance
somewhat.