vga: fix flash of garbage when band list is cleared.
Previously, if we got to the end of the vertical blanking interval with
a null band list, we'd suppress rasterization...but do so in a way that
caused whatever happened to be in the scan buffer to be displayed.

Which is crappy.

Now, we configure scanout for a zero-pixel line, which effectively
suppresses video.

For best results, don't hit this case: replace the band list before it's
needed.